+++ backend/models/foodIngredients.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
     const FoodIngredient = sequelize.define('FoodIngredient', {
         mealId: DataTypes.INTEGER,
         ingredientId: DataTypes.INTEGER,
-        weight: DataTypes.FLOAT
+        weight: DataTypes.FLOAT(2)
     }, {});
     FoodIngredient.associate = function(models) {
