+++ backend/models/meal.js
@@ -2,10 +2,10 @@
 module.exports = (sequelize, DataTypes) => {
   const Meal = sequelize.define('Meal', {
     name: DataTypes.STRING,
-    kcal: DataTypes.FLOAT,
-    proteins: DataTypes.FLOAT,
-    carbs: DataTypes.FLOAT,
-    fats: DataTypes.FLOAT
+    kcal: DataTypes.FLOAT(2),
+    proteins: DataTypes.FLOAT(2),
+    carbs: DataTypes.FLOAT(2),
+    fats: DataTypes.FLOAT(2)
   }, {});
   Meal.associate = function(models) {
     Meal.belongsToMany(models.Ingredient, { through: models.FoodIngredient });
